+++ scripts/action-forge.js
@@ -22,7 +22,7 @@ Hooks.once("init", () => {
     });
   }
 
-  console.info(`PF2E Action Forge | Initialized ${module?.version ?? "0.1.0-dev.7.2"}`);
+  console.info(`PF2E Action Forge | Initialized ${module?.version ?? "0.1.0-dev.7.3"}`);
 });
 
 Hooks.on("getSceneControlButtons", (controls) => {